requirements.txt
google-cloud-aiplatform>=1.60.0
google-adk>=0.5.0
playwright==1.44.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.16.0
pydantic>=2.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
nest_asyncio>=1.6.0
